server/requirements.txt
arlo==1.2.35
astroid==2.4.2
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
-e git+https://github.com/OxfordHCC/social_translucence_cams.git@7afb31802de843fa8ac50d479f6b222dde962ae2#egg=fake_arlo&subdirectory=server/fake_arlo
Flask==1.1.2
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
idna==2.10
iniconfig==1.0.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
more-itertools==8.4.0
packaging==20.4
peewee==3.13.3
pluggy==0.13.1
psutil==5.7.2
py==1.9.0
PyJWT==1.7.1
pylint==2.5.3
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
pytest==6.0.1
pytest-xprocess==0.13.1
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
sseclient==0.0.22
toml==0.10.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
